RouteList: module handling extracted to ModuleRouter (BC break)
