+++ src/Application/Routers/RouteList.php
@@ -14,21 +14,11 @@ use Nette;
  * The router broker.
  *
  * @author     David Grudl
- * @property-read string $module
  */
 class RouteList extends Nette\Utils\ArrayList implements Nette\Application\IRouter
 {
 	/** @var array */
 	private $cachedRoutes;
-
-	/** @var string */
-	private $module;
-
-
-	public function __construct($module = NULL)
-	{
-		$this->module = $module ? $module . ':' : '';
-	}
 
 
 	/**
@@ -40,10 +30,6 @@ class RouteList extends Nette\Utils\ArrayList implements Nette\Application\IRout
 		foreach ($this as $route) {
 			$appRequest = $route->match($httpRequest);
 			if ($appRequest !== NULL) {
-				$name = $appRequest->getPresenterName();
-				if (strncmp($name, 'Nette:', 6)) {
-					$appRequest->setPresenterName($this->module . $name);
-				}
 				return $appRequest;
 			}
 		}
@@ -85,15 +71,6 @@ class RouteList extends Nette\Utils\ArrayList implements Nette\Application\IRout
 			$this->cachedRoutes = $routes;
 		}
 
-		if ($this->module) {
-			if (strncasecmp($tmp = $appRequest->getPresenterName(), $this->module, strlen($this->module)) === 0) {
-				$appRequest = clone $appRequest;
-				$appRequest->setPresenterName(substr($tmp, strlen($this->module)));
-			} else {
-				return NULL;
-			}
-		}
-
 		$presenter = strtolower($appRequest->getPresenterName());
 		if (!isset($this->cachedRoutes[$presenter])) {
 			$presenter = '*';
@@ -122,15 +99,6 @@ class RouteList extends Nette\Utils\ArrayList implements Nette\Application\IRout
 			throw new Nette\InvalidArgumentException('Argument must be IRouter descendant.');
 		}
 		parent::offsetSet($index, $route);
-	}
-
-
-	/**
-	 * @return string
-	 */
-	public function getModule()
-	{
-		return $this->module;
 	}
 
 }
